unmask SIGCHILD so waitpid won't return -ECHILD
the OpenVPN Installer that replaces the openvpn binary in /system/xbin also won't help.

I guess something changed in the framework that starts binaries with SIGCHILD
masked (the apk that spawn the openvpn binary hasn't changed).

This leads to a failure in waitpid that returns -ECHILD and openvpn instead
expects the pid.

No idea why SIGCHILD it's masked and where... but this is a recent regression.

with this fix the OpenVPN Installer binary still won't work of course, but
the stock cm10 binary will work fine and that's the important thing.

it should be investigated if SIGCHILD masked is a feature of a bug.

Change-Id: I1fcfa2666ca086b1bc62659692fc74bad16cc0cc